refactor(file-parsers): Move module from common to electron
The file parsers module relies on Node.js's `fs` API, making it unsuitable for common environments like browsers.
To clarify its environmental dependency, it has been refactored and moved under the `@talex-touch/utils/electron` package.

BREAKING CHANGE: The import path for the file parsers module has changed from '@talex-touch/utils/common/file-parsers' to '@talex-touch/utils/electron/file-parsers'.
