+++ packages/utils/common/index.ts
@@ -1,4 +1,4 @@
 export * from './storage/index'
 export * from './utils'
 export * from './search'
-export * from './file-parsers'
+export * from '../electron/file-parsers'
