+++ packages/utils/electron/file-parsers/parsers/text-parser.ts
