+++ packages/utils/electron/file-parsers/registry.ts
