+++ packages/utils/electron/file-parsers/types.ts
